+++ boluo/src/lib.rs
@@ -266,7 +266,7 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg, doc_cfg))]
 
 pub use boluo_core::BoxError;
-pub use boluo_core::{body, handler, http, request, service};
+pub use boluo_core::{body, handler, http, request, service, upgrade};
 
 pub use boluo_macros::route;