+++ boluo/CHANGELOG.md
@@ -1,5 +1,7 @@
 # unreleased
 
+# 0.6.2
+
 ## 变化
 
 - `Router` 的 `merge`，`merge_with`，`try_merge`，`try_merge_with` 方法现在接受任何可以转换为 `Router` 的类型。